Provide basis of alternative client
This client will have nicer user interface. The client allows loading
plugins from predefined directories that implement actual subcommands or
enhance them (so that implementing internal extensions is possible). No
actual commands are implemented in this patch.

Each plugin implementing some commands must be derived from provided
parent class. This parent has helper methods for adding commands. These
commands can be defined to not be shown in help output by default
(unless requested with --help-all).

There is a helper class PDCClient that allows user to instantiate
BeanBag by just providing a URL or shortcut from settings. It handles
obtaining the token transparently.

The client tries to present errors to user in a readable way. It does
not handle all possible error conditions, though.

The client supports Bash autocompletion, assuming argcomplete Python
package is available. It is not added to requirements, as it makes not
sense to install it on the server and it is not a hard dependency
anyway.

JIRA: PDC-630