Add RPM plugin to client
This plugin implements commands for listing, showing, creating and
updating RPM metadata in PDC. When retrieving information about
particular RPM (or when updating), the database identifier is used. This
could be replaced by NEVRA in the future, but it would require
dependency on a library that would parse it.

JIRA: PDC-1023