Move get_paged function to pdc_client
This function is useful for all client applications, not just pdc
plugins. Now anything using PDCClient class can use this helper.

This is a backwards incompatible change. Since the client has not been
released yet, it's not a big deal. Existing plugins are updated to
import the function from its new place.