Fix SSL command line options
The behavior of `-k`/`--insecure` and `--ca-cert` is incorrect with no
configuration. The concept of "insecure" should also be restricted to
the command line parser; the client object should only know the value to
hand off to `requests`.

This patch corrects this issue, as well as another issue where command
line arguments would not override configuration values. The basic client
(`pdc_client` script) has been switched from `OptionParser` to
`ArgumentParser` to simplify and keep synchronized with the rich client.