Simplify parsing responses
Instead of getting text from the response and calling json.loads
ourselves, we can directly ask requests to do the parsing for us.

The advantage to this approach is that encoding detection works a little
better. Per JSON RFC 4627 section 3 some variant of UTF should be used,
but the `.text` attribute tries to guess using chardet which may come up
with something else.